log: add caller plugin-id about emitting error events
It would be helpful if we could know what plugin emitted the
error event.

https://github.com/fluent/fluentd/issues/4567

We need to care about the compatibility.
This signature change would not break compatibility.

However, I'm concerned that `caller_plugin_id` has a race
condition, although I don't confirm it.
It looks to me that the id can be another plugin-id running
concurrently...
It is not the issue with this fix, it is the issue of the
existing implementation.

Signed-off-by: Daijiro Fukuda <fukuda@clear-code.com>